fix: Repository 키 Long 으로 수정

+++ server/src/main/java/com/seniclass/server/domain/lecture/repository/WidgetSettingRepository.java
@@ -4,7 +4,10 @@ import com.seniclass.server.domain.lecture.domain.WidgetSetting;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.Optional;
+
 @Repository
-public interface WidgetSettingRepository extends JpaRepository<WidgetSetting, Integer> {
+public interface WidgetSettingRepository extends JpaRepository<WidgetSetting, Long> {
     boolean existsByLectureId(Long lectureId);
 }
